重構: 整理假資料檔案
