+++ frontend/src/views/Admin/components/MemberDetailCarOwner.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { computed } from 'vue';
-import type { Member } from '@/views/Admin/composables/useMemberData';
+import type { Member } from '@/views/Admin/composables/memberCore';
 import { getStatusLabel, getStatusClass } from '../composables/memberCore';
 
 const props = defineProps<{
